devops

+++ user-service/src/main/java/com/example/userservice/service/UserSyncService.java
@@ -67,7 +67,7 @@ public class UserSyncService {
             try {
                 if (attempt > 1) {
                     // Délai exponentiel : 2s, 4s, 8s, 16s, 32s
-                    long delay = (long) (INITIAL_RETRY_DELAY_MS * Math.pow(2.0, attempt - 2));
+                    long delay = (long) (INITIAL_RETRY_DELAY_MS * Math.pow(2.0, (double) (attempt - 2)));
                     log.info("Attente de {} secondes avant la tentative {} de récupération de l'utilisateur Keycloak", delay/1000, attempt);
                     Thread.sleep(delay);
                 }
